Refactor vector test utils to a common class.
